+++ packages/hooks/src/useConfig.ts
@@ -6,23 +6,30 @@ import { Config } from './graphql'
 type Result<T = any> = {
   data?: T
   error?: Error
+  loading: boolean
 }
 
-export default function useConfig(): Result<Config> {
+export type ConfigResult = Result<Config>
+
+export default function useConfig(): ConfigResult {
   const { sdk } = useContext(LiteflowContext)
   const [config, setConfig] = useState<Config>()
   const [error, setError] = useState<Error>()
+  const [loading, setLoading] = useState<boolean>(false)
 
   useEffect(() => {
+    setLoading(true)
     sdk
       .GetConfig()
-      .then(({ config }) => config)
+      .then(({ config }) => setConfig(config))
       .catch(setError)
+      .finally(() => setLoading(false))
     return () => setConfig(undefined)
   }, [setConfig, setError, sdk])
 
   return {
     data: config,
     error,
+    loading,
   }
 }